+++ benchmarks/src/CMakeLists.txt
@@ -60,6 +60,8 @@ target_compile_options(SerializationTwitterBenchmark PRIVATE -freflection -stdli
 # Link sanitizers and libraries
 target_link_libraries(SerializationTwitterBenchmark PRIVATE nlohmann_json::nlohmann_json simdjson::simdjson )
 
+target_link_libraries(SerializationBenchmark PRIVATE simdjson-serial nlohmann_json::nlohmann_json simdjson::simdjson)
+
 # cpp-reflect may fail to build on LLVM 19 due to this issue
 # https://github.com/hanickadot/compile-time-regular-expressions/issues/307
 option(SIMDJSON_BENCH_CPP_REFLECT "include cpp reflect in the benchmarks" OFF)
@@ -67,12 +69,18 @@ if(SIMDJSON_BENCH_CPP_REFLECT)
   message(STATUS "Including reflect-cpp in the benchmarks.")
   add_library(benchmark_reflect_serialization_twitter STATIC benchmark_reflect_serialization_twitter.cpp)
   target_link_libraries(benchmark_reflect_serialization_twitter PRIVATE reflectcpp)
+
   target_link_libraries(SerializationTwitterBenchmark PRIVATE benchmark_reflect_serialization_twitter)
-  target_compile_definitions(benchmark_reflect_serialization_twitter PRIVATE SIMDJSON_BENCH_CPP_REFLECT=1)
+  target_compile_definitions(SerializationTwitterBenchmark PRIVATE SIMDJSON_BENCH_CPP_REFLECT=1)
+
+
+  add_library(benchmark_reflect_serialization STATIC benchmark_reflect_serialization_twitter.cpp)
+  target_link_libraries(benchmark_reflect_serialization PRIVATE reflectcpp)
+
+  target_link_libraries(SerializationBenchmark PRIVATE benchmark_reflect_serialization)
+  target_compile_definitions(SerializationBenchmark PRIVATE SIMDJSON_BENCH_CPP_REFLECT=1)
 endif()
 
-
-target_link_libraries(SerializationBenchmark PRIVATE simdjson-serial nlohmann_json::nlohmann_json simdjson::simdjson)
 
 
 target_compile_definitions(SerializationTwitterBenchmark PRIVATE JSON_FILE="${CMAKE_CURRENT_SOURCE_DIR}/data/twitter.json")
