+++ benchmarks/src/benchmark_reflect_serialization.hpp
@@ -1,0 +1,6 @@
+#ifndef BENCHMARK_REFLECT_SERIALIZATION_TWITTER_HPP
+#define BENCHMARK_REFLECT_SERIALIZATION_TWITTER_HPP
+
+#include "user_profile.hpp"
+void bench_reflect_cpp(std::vector<User>& data)
+#endif
